pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "searchjobs"
version = "0.1.0"
description = "A multi-agent job search and application system using A2A protocol."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Ranjan Desai", email = "ranjan@hubhaya.com" }
]
dependencies = [
    "a2a-sdk==0.3.2",
    "google-adk==1.25.0",
    "ddgs==8.1.1",
    "uvicorn==0.40.0",
    "httpx==0.28.1",
    "starlette==0.52.1",
    "pydantic==2.12.5",
    "google-genai==1.63.0",
    "sse-starlette==3.2.0",
    "python-dotenv==1.2.1",
    "fastapi==0.128.8",
]

[project.scripts]
search-agent = "search_agent.__main__:main"
apply-agent = "apply_agent.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["search_agent", "apply_agent"]
